# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Fake News Detection System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface PredictInput {
  /**
     * News article or headline to analyze
     * @minLength 10
     */
  text: string;
}

export type PredictionResultLabel = typeof PredictionResultLabel[keyof typeof PredictionResultLabel];


export const PredictionResultLabel = {
  REAL: 'REAL',
  FAKE: 'FAKE',
} as const;

export type KeyWordSentiment = typeof KeyWordSentiment[keyof typeof KeyWordSentiment];


export const KeyWordSentiment = {
  positive: 'positive',
  negative: 'negative',
  neutral: 'neutral',
} as const;

export interface KeyWord {
  word: string;
  /** Influence score (0-1) */
  score: number;
  sentiment: KeyWordSentiment;
}

export interface PredictionResult {
  label: PredictionResultLabel;
  /** Confidence percentage (0-100) */
  confidence: number;
  /** Probability of being real (0-1) */
  realProbability: number;
  /** Probability of being fake (0-1) */
  fakeProbability: number;
  /** Time taken to process in milliseconds */
  processingTimeMs: number;
  /** Human-readable explanation of the prediction */
  explanation: string;
  keyWords: KeyWord[];
  /** Whether the result was produced by the AI model (true) or the rule-based fallback (false) */
  aiPowered?: boolean;
}

export type PredictionRecordLabel = typeof PredictionRecordLabel[keyof typeof PredictionRecordLabel];


export const PredictionRecordLabel = {
  REAL: 'REAL',
  FAKE: 'FAKE',
} as const;

export interface PredictionRecord {
  id: number;
  text: string;
  label: PredictionRecordLabel;
  confidence: number;
  createdAt: string;
}

export interface DashboardStats {
  datasetSize: number;
  realCount: number;
  fakeCount: number;
  modelName: string;
  accuracy: number;
  precision: number;
  recall: number;
  f1Score: number;
  totalPredictions: number;
}

export interface ClassDistribution {
  real: number;
  fake: number;
}

export interface ModelScore {
  name: string;
  accuracy: number;
  precision: number;
  recall: number;
  f1Score: number;
}

export interface ConfusionMatrix {
  truePositive: number;
  trueNegative: number;
  falsePositive: number;
  falseNegative: number;
}

export interface ErrorResponse {
  error: string;
}

export type GetPredictionHistoryParams = {
limit?: number;
};