# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Fake News Detection System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /predict:
    post:
      operationId: predictNews
      tags: [predict]
      summary: Predict if news is real or fake
      description: Analyzes a news article or headline and returns a real/fake prediction with confidence and explanation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PredictInput"
      responses:
        "200":
          description: Prediction result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PredictionResult"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /predictions/history:
    get:
      operationId: getPredictionHistory
      tags: [predict]
      summary: Get recent prediction history
      description: Returns recent predictions made by users
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of past predictions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PredictionRecord"
...
components:
  schemas:
...
    PredictInput:
      type: object
      required:
        - text
      properties:
        text:
          type: string
          minLength: 10
          description: News article or headline to analyze
    PredictionResult:
      type: object
      required:
        - label
        - confidence
        - realProbability
        - fakeProbability
        - processingTimeMs
        - explanation
        - keyWords
      properties:
        label:
          type: string
          enum: [REAL, FAKE]
        confidence:
          type: number
          description: Confidence percentage (0-100)
        realProbability:
          type: number
          description: Probability of being real (0-1)
        fakeProbability:
          type: number
          description: Probability of being fake (0-1)
        processingTimeMs:
          type: number
          description: Time taken to process in milliseconds
        explanation:
          type: string
          description: Human-readable explanation of the prediction
        keyWords:
          type: array
          items:
            $ref: "#/components/schemas/KeyWord"
        aiPowered:
          type: boolean
          description: Whether the result was produced by the AI model (true) or the rule-based fallback (false)
    KeyWord:
      type: object
      required:
        - word
        - score
        - sentiment
      properties:
        word:
          type: string
        score:
          type: number
          description: Influence score (0-1)
        sentiment:
          type: string
          enum: [positive, negative, neutral]
    PredictionRecord:
      type: object
      required:
        - id
        - text
        - label
        - confidence
        - createdAt
      properties:
        id:
          type: integer
        text:
          type: string
        label:
          type: string
          enum: [REAL, FAKE]
        confidence:
          type: number
        createdAt:
          type: string
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string