# lib/api-zod/src/generated/types/keyWord.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Fake News Detection System API
 * OpenAPI spec version: 0.1.0
 */
import type { KeyWordSentiment } from './keyWordSentiment';

export interface KeyWord {
  word: string;
  /** Influence score (0-1) */
  score: number;
  sentiment: KeyWordSentiment;
}