# lib/api-zod/src/generated/types/predictInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Fake News Detection System API
 * OpenAPI spec version: 0.1.0
 */

export interface PredictInput {
  /**
     * News article or headline to analyze
     * @minLength 10
     */
  text: string;
}