# lib/api-zod/src/generated/types/predictionRecord.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Fake News Detection System API
 * OpenAPI spec version: 0.1.0
 */
import type { PredictionRecordLabel } from './predictionRecordLabel';

export interface PredictionRecord {
  id: number;
  text: string;
  label: PredictionRecordLabel;
  confidence: number;
  createdAt: string;
}