# lib/api-zod/src/generated/types/predictionResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Fake News Detection System API
 * OpenAPI spec version: 0.1.0
 */
import type { KeyWord } from './keyWord';
import type { PredictionResultLabel } from './predictionResultLabel';

export interface PredictionResult {
  label: PredictionResultLabel;
  /** Confidence percentage (0-100) */
  confidence: number;
  /** Probability of being real (0-1) */
  realProbability: number;
  /** Probability of being fake (0-1) */
  fakeProbability: number;
  /** Time taken to process in milliseconds */
  processingTimeMs: number;
  /** Human-readable explanation of the prediction */
  explanation: string;
  keyWords: KeyWord[];
  /** Whether the result was produced by the AI model (true) or the rule-based fallback (false) */
  aiPowered?: boolean;
}